refactor: Minor refactor on the resolve function
