+++ crates/bld_runner/src/expr/v3/mod.rs
@@ -1,5 +1,4 @@
 pub mod context;
 pub mod exec;
 pub mod parser;
-pub mod startup;
 pub mod traits;
